requirements.txt
disco-py==0.0.12
pymongo==3.7.1
matplotlib==3.0.0